requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
pymongo==3.4.0
requests==2.18.2
stem==1.5.4
urllib3==1.22
eventlet==0.25.0
Flask-SocketIO==4.1.0